move handler funtionality to SSHSubsystemFwup module
